+++ EleFits/EleFits/MefFile.h
@@ -17,9 +17,11 @@
 namespace Euclid {
 namespace Fits {
 
+/// @cond
 // Forward declaration for MefFile::filter()
-template <typename THdu>
+template <typename T>
 class HduSelector;
+/// @endcond
 
 /**
  * @ingroup file_handlers
@@ -31,15 +33,15 @@ class HduSelector;
  * image HDUs and binary table HDUs (see \ref primer).
  * Methods to access HDUs, header units or data units all return constant references,
  * because they are stateless views on the `MefFile` object,
- * which is the only handler modified by write operation.
+ * which is the only handler modified by write operations.
  * 
  * At creation, a `MefFile` already contains a Primary,
  * which is empty but can be resized and filled.
- * HDUs are accessed either directly by their index, e.g. with `access()`,
+ * HDUs are accessed either directly by their index, e.g. with `access()` or operator `[]`,
  * or by finding a {type, name, version} triplet with `find()`.
  * Although it should not be an issue for most files (even with hundreds of HDUs),
  * the second option is much slower because it consists in reading each header unit until a match is found.
- * Once a HDU has been accessed, the reference can be safely reused
+ * Once an HDU has been accessed, the reference can be safely reused
  * (please don't re-find an already accessed HDU).
  * It is possible to specify the return type of `access()` and `find()` to best fit target usage, e.g.:
  * 
@@ -90,7 +92,7 @@ public:
   /**
    * @brief Get the number of HDUs.
    * 
-   * As opposed to CFITSIO's HDU counting, the number is not read by this function:
+   * As opposed to CFITSIO's HDU counting, incomplete HDUs are taken into account:
    * it is initialized by the constructor and then updated at each modification through `MefFile` methods.
    * This way, incomplete HDUs are also taken into account where CFITSIO would exclude them.
    * This means, for example, that the initial number of HDUs in a new file is 1 instead of 0 with CFITSIO.
@@ -122,11 +124,11 @@ public:
   /**
    * @brief Access the HDU at given 0-based index.
    * @tparam T The type of HDU or header or data unit handler, i.e.
-   * `ImageHdu`, `ImageRaster`, `BintableHdu`, `BintableColumns`, `Hdu` or `Header` to just handle metadata.
+   * `ImageHdu`, `ImageRaster`, `BintableHdu`, `BintableColumns`, `Hdu` or `Header`.
    * @return A reference to the HDU reader-writer.
    * 
    * Backward indexing is enabled.
-   * The default handler is `Hdu`, in which case the returned HDU can still be cast to another handler, e.g.:
+   * The default handler is `Hdu`, in which case the returned HDU can still be cast to another handler with `Hdu::as()`, e.g.:
    * \code
    * const auto &ext = f.access<>(-1); // Same as f[-1]
    * auto raster = ext.as<ImageHdu>().raster().read<float>();
@@ -140,26 +142,27 @@ public:
   const T& access(long index);
 
   /**
-   * @brief Shortcut for access<Hdu>(long)
+   * @brief Shortcut for `access<Hdu>(long)`
    */
   const Hdu& operator[](long index);
 
   /**
    * @brief Access the first HDU with given name, type and version.
-   * @tparam T The type of HDU or data unit
+   * @tparam T The type of HDU or header or data unit handler
    * @param name The HDU name
    * @param version The HDU version, or 0 to not check the version
    * 
    * The template parameter is used to disambiguate when two extensions of different types have the same name.
    * For example, in a file with an image extension and a binary table extension both named "EXT",
-   * `find<ImageHdu>("EXT")` returns the image extension,
+   * `find<ImageHdu>("EXT")` or `find<ImageRaster>()` return the image extension,
    * while `find<BintableHdu>("EXT")` returns the binary table extension,
    * and `find<Hdu>("EXT")` returns whichever of the two has the smallest index.
    * 
-   * In the case where several HDUs of same type have the same type, name and version
+   * In the case where several HDUs have the same type, name and version
    * (which is discouraged by the standard, but not forbidden),
    * method `readHduNames()` should be used to get the indices
    * and then `access(long)` should be called.
+   * 
    * @see access(long)
    * @see access(const std::string&)
    */
@@ -168,11 +171,14 @@ public:
 
   /**
    * @brief Access the only HDU with given name, type and version.
+   * @tparam T The type of HDU or header or data unit handler
    * 
    * Throws an exception if several HDUs with given name exists.
+   * 
    * @warning
    * In order to ensure uniqueness of the name, all HDUs are visited,
    * which may have a non-negligible I/O cost.
+   * 
    * @see find()
    */
   template <class T = Hdu>
@@ -188,6 +194,7 @@ public:
    * @ingroup iterators
    * @brief Get an iterable object which represents a filtered set of HDUs.
    * @return An iterable object, i.e. one for which `begin()` and `end()` functions are provided.
+   * 
    * @par_example
    * \code
    * for (const auto& hdu : f.filter<ImageHdu>(HduCategory::Created)) {
@@ -196,25 +203,35 @@ public:
    * \endcode
    * @see `iterators`
    */
-  template <typename THdu = Hdu>
-  HduSelector<THdu> filter(const HduFilter& categories = HduCategory::Any);
+  template <typename T = Hdu>
+  HduSelector<T> filter(const HduFilter& categories = HduCategory::Any);
 
   /// @group_modifiers
 
   /**
    * @brief Set the compression algorithm to be used when writting new image extensions.
    * 
+   * From the call on, all new image extensions are created as compressed image extensions.
+   * Compression can then be turned off with `stopCompressing()`.
+   * 
    * @warning If any, the compression strategy is destroyed by this method.
+   * 
+   * @see CompressionStrategy
    */
   inline void startCompressing(const Compression& algo);
 
   /**
    * @brief Set the compression strategy to be used when writting new image extensions.
+   * 
+   * From the call on, all new image extensions are created as compressed image extensions,
+   * where the compression algorithm is deduced from the extension properties (type, size, records...).
+   * 
+   * Compression can be turned off with `stopCompressing()`.
    */
   inline void startCompressing(std::unique_ptr<CompressionStrategy> strategy);
 
   /**
-   * @brief Disable compression when writting new image extensions.
+   * @brief Disable compression.
    */
   inline void stopCompressing();
 
@@ -225,13 +242,20 @@ public:
   const ImageHdu& appendImageHeader(const std::string& name = "", const RecordSeq& records = {});
 
   /**
-   * @brief Append and new image extension and fill the data unit with null values.
+   * @brief Append a new image extension and fill the data unit with null values.
+   * @tparam T The value type
+   * @tparam N The dimension parameter
    * 
    * The null value is defined as follows:
    * - For integer images, if the `BLANK` record is among the provided ones,
    *   then its value is used unscaled (`BZERO` is not taken into account);
    *   Otherwise, the data unit is filled with zeros.
    * - For real images, the data unit is filled with NaNs.
+   * 
+   * @par_example
+   * \code
+   * f.appendNullImage<char, 3>("CUBE", {{"BLANK", 255}}, {28, 06, 1989});
+   * \endcode
    */
   template <typename T, long N = 2>
   const ImageHdu& appendNullImage(const std::string& name, const RecordSeq& records, const Position<N>& shape);
@@ -248,17 +272,18 @@ public:
   /**
   * @brief Append a copy of a given HDU.
   * 
-  * If the copied HDU is an Image HDU, the current compression algorithm of the `MefFile` is applied.
-  * If the copied HDU is a Primary, as its appended as an Image HDU, and current compression is also applied.
+  * If the copied HDU is an image HDU larger than a block of data (2880 bytes),
+  * then the current compression algorithm of the `MefFile` is applied.
+  * Additionally, if the copied HDU is a Primary, it is appended as an image extension.
   * 
-  * @warning Does NOT support copying the HDU to the SAME MefFile as the one it originates from.
-  * @warning When copying an empty image, compression is NEVER applied.
+  * @warning
+  * The source and destination `MefFile`s must be different.
   */
-  template <typename THdu = Hdu>
-  const THdu& appendCopy(const THdu& hdu);
+  template <typename T = Hdu>
+  const T& appendCopy(const T& hdu);
 
   /**
-   * @brief Append a binary table extension with empty data unit (0 rows and possibly 0 columns).
+   * @brief Append a binary table extension with empty data unit (0 rows, and possibly 0 columns).
    */
   template <typename... TInfos>
   const BintableHdu&
@@ -281,8 +306,7 @@ public:
    * @brief Append and write a new binary table extension.
    */
   template <typename... TColumns>
-  const BintableHdu&
-  appendBintable(const std::string& name = "", const RecordSeq& records = {}, const TColumns&... columns);
+  const BintableHdu& appendBintable(const std::string& name, const RecordSeq& records, const TColumns&... columns);
 
   /**
    * @brief Append and write a new binary table extension.
@@ -291,99 +315,6 @@ public:
   const BintableHdu& appendBintable(const std::string& name, const RecordSeq& records, const TColumns& columns);
 
   /// @}
-  /// @name Deprecated
-  /// @{
-
-  /**
-   * @brief Deprecated alias for backward compatibility.
-   * @deprecated see find()
-   */
-  template <class T = Hdu>
-  const T& accessFirst(const std::string& name, long version = 0) {
-    return find<T>(name, version);
-  }
-
-  /**
-   * @brief Deprecated alias for backward compatibility.
-   * @deprecated see filter()
-   */
-  template <typename THdu = Hdu>
-  HduSelector<THdu> select(const HduFilter& categories = HduCategory::Any) {
-    return filter(categories);
-  }
-
-  /**
-   * @brief Append a new Hdu (as an empty ImageHdu) with given name.
-   * @return A reference to the new Hdu.
-   * @deprecated See appendImageHeader
-   */
-  const Hdu& initRecordExt(const std::string& name);
-
-  /**
-   * @brief Append a new ImageHdu with given name and shape.
-   * 
-   * To not only initialize the HDU but also write data, use assignImageExt instead.
-   * @deprecated See appendImageExt
-   */
-  template <typename T, long N>
-  const ImageHdu& initImageExt(const std::string& name, const Position<N>& shape);
-
-  /**
-   * @brief Append an ImageHdu with given name and data.
-   * @return A reference to the new ImageHdu.
-   * @deprecated See appendImageExt
-   */
-  template <typename TRaster>
-  const ImageHdu& assignImageExt(const std::string& name, const TRaster& raster);
-
-  /**
-   * @brief Append a BintableHdu with given name and columns info.
-   * 
-   * To not only initialize the HDU but also write data, use assignBintableExt instead.
-   * @deprecated See appendBintableHeader
-   */
-  template <typename... TInfos>
-  const BintableHdu& initBintableExt(const std::string& name, const TInfos&... header);
-
-  /**
-   * @brief Append a BintableHdu with given name and data.
-   * @return A reference to the new BintableHdu.
-   * @warning
-   * All columns should have the same number of rows.
-   * @deprecated See appendBintableExt
-   */
-  template <typename... TColumns>
-  const BintableHdu& assignBintableExt(const std::string& name, const TColumns&... columns);
-
-  /**
-   * @brief Append a BintableHdu with given name and data.
-   * @return A reference to the new BintableHdu.
-   * @tparam TTuple A tuple of columns
-   * @tparam Size The number of columns
-   * @warning
-   * All columns should have the same number of rows.
-   * @deprecated See appendBintableExt
-   */
-  template <typename TTuple, std::size_t Size = std::tuple_size<TTuple>::value>
-  const BintableHdu& assignBintableExt(const std::string& name, const TTuple& columns);
-
-  /// @}
-
-  /**
-   * @brief The index of the Primary HDU.
-   * 
-   * The HDU indices are now 0-based while they used to be 1-based.
-   * This constant makes migration smoother.
-   * It can be used to loop safely over the HDUs:
-   * \code
-   * for (long i = MefFile::primaryIndex; i < MefFile::primaryIndex + f.hduCount(); ++i) {
-   *   const auto &ext = f.access<>(i);
-   *   // do something with ext
-   * }
-   * \endcode
-   * @deprecated No more useful now that the indexing change is old.
-   */
-  static constexpr long primaryIndex = 0;
 
 protected:
   /**